Updating scripts

+++ post-zsh.sh
@@ -1,21 +1,48 @@
-git clone https://github.com/zsh-users/zsh-autosuggestions.git $ZSH_CUSTOM/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $ZSH_CUSTOM/plugins/zsh-syntax-highlighting
-git clone https://github.com/denysdovhan/spaceship-prompt.git "$ZSH_CUSTOM/themes/spaceship-prompt"
-ln -s "$ZSH_CUSTOM/themes/spaceship-prompt/spaceship.zsh-theme" "$ZSH_CUSTOM/themes/spaceship.zsh-theme"
+#!/bin/zsh
 
-echo -e "\U1f9d1\U200d\U1f4bb \e[1;33mCustomizing .zshrc...\e[0m"
-rm ~/.zshrc
-curl -o ~/.zshrc https://raw.githubusercontent.com/jeremyfuksa/standard-pi-setup/main/zshrc
+# Function to display status messages
+show_status() {
+		echo "$(tput setaf 4)>>> $1$(tput sgr0)"
+}
+
+# Clone or update Oh My Zsh plugins and Spaceship theme
+clone_or_update_plugin() {
+		local repo_url="$1"
+		local destination="$2"
+
+		if [[ ! -d "$destination" ]]; then
+				show_status "Cloning $destination..."
+				git clone "$repo_url" "$destination"
+		else
+				show_status "$destination already exists. Updating..."
+				(cd "$destination" && git pull origin)
+		fi
+}
+
+# Install custom .zshrc
+show_status "Installing custom .zshrc..."
+curl -o ~/.zshrc https://gist.githubusercontent.com/jeremyfuksa/440f27c77537c3d2382431f589506e8e/raw/4a4cb841c285d12e6634761b5c50684685a89cc3/.zshrc
 source ~/.zshrc
 
-echo -e "\U1f9d1\U200d\U1f4bb \e[1;33mInstalling .zprofile...\e[0m"
-curl -o ~/.zprofile https://raw.githubusercontent.com/jeremyfuksa/standard-pi-setup/main/zprofile
-source ~/.zprofile
+# Clear stock MOTD
+show_status "Clearing stock MOTD..."
+echo -n > /etc/motd
 
-echo -e "\U1f9d1\U200d\U1f4bb \e[1;33mConfiguring git...\e[0m"
-git config --global pull.rebase false
+# Clone or update zsh-autosuggestions
+clone_or_update_plugin "https://github.com/zsh-users/zsh-autosuggestions.git" "$ZSH_CUSTOM/plugins/zsh-autosuggestions"
 
-echo -e "\U1f9d1\U200d\U1f4bb \e[1;33mGetting the latest OS updates...\e[0m"
-update-all
+# Clone or update zsh-syntax-highlighting
+clone_or_update_plugin "https://github.com/zsh-users/zsh-syntax-highlighting.git" "$ZSH_CUSTOM/plugins/zsh-syntax-highlighting"
 
-echo -e "\n\U1f967 \e[1;32mThis Pi is sliced! \e[0m"
+# Clone or update spaceship-prompt theme and create symlink for spaceship.zsh-theme
+clone_or_update_plugin "https://github.com/denysdovhan/spaceship-prompt.git" "$ZSH_CUSTOM/themes/spaceship-prompt"
+if [[ ! -e "$ZSH_CUSTOM/themes/spaceship.zsh-theme" ]]; then
+		show_status "Creating symlink for spaceship.zsh-theme..."
+		ln -s "$ZSH_CUSTOM/themes/spaceship-prompt/spaceship.zsh-theme" "$ZSH_CUSTOM/themes/spaceship.zsh-theme"
+else
+		show_status "spaceship.zsh-theme symlink already exists."
+fi
+
+# Run the update-all command in zsh
+show_status "Running update-all command in zsh..."
+zsh -c "update-all"
